+++ src/builder.ts
@@ -67,7 +67,8 @@ import {
 } from './consts'
 import { unknownParsedFile } from './apitypes/unknown/unknown.parser'
 import { createVersionPackage } from './components/package'
-import { applyBuilderVersionInfo, compareVersions } from './components/compare'
+import { compareVersions } from './components/compare'
+import { applyBuilderVersionInfo } from './validators'
 import { buildFiles } from './components/files'
 import JSZip from 'jszip'
 import { calculateHistoryForDeprecatedItems } from './components/deprecated'
